Improve the operation of the upload functionality.
Now, uploads are done in three phases, two interactive, and one
automatic. From index.n/upload the user chooses the xml file to
upload, then the system analyses it, and presents a list of
results that will be generated along with information about any
non-existant configs, tags or tests.

The xml can now include extra elements to define new confits, tags
and tests. There is no support to specify the config in the uload
ui now, it must be specified in the xml.

If a test with the config, tag and test id already exists in the
database, the confirmation list will higlight it, and the time
will be updated, instead of adding a new result.